Refactor: rename listener package to 'event'
