+++ core/src/main/java/de/pxav/kelp/core/event/ConditionalExpiryTestStage.java
@@ -1,4 +1,4 @@
-package de.pxav.kelp.core.listener;
+package de.pxav.kelp.core.event;
 
 /**
  * This enum says in which execution stage a conditional expiry
