+++ core/src/main/java/de/pxav/kelp/core/event/EventHandlerRegistration.java
@@ -1,4 +1,4 @@
-package de.pxav.kelp.core.listener;
+package de.pxav.kelp.core.event;
 
 import com.google.common.base.Preconditions;
 import com.google.inject.Inject;
