+++ core/src/main/java/de/pxav/kelp/core/event/Subscribes.java
@@ -1,4 +1,4 @@
-package de.pxav.kelp.core.listener;
+package de.pxav.kelp.core.event;
 
 import org.bukkit.event.Event;
